# MAINT: optimization and broadcasting for .replace() method for strings.