--- a/numpy/_core/defchararray.py
+++ b/numpy/_core/defchararray.py
@@ -1357,16 +1357,16 @@ def replace(a, old, new, count=None):
     ----------
     a : array-like of str or unicode
 
-    old, new : str or unicode
+    old, new : scalar or array-like str or unicode
 
-    count : int, optional
+    count : scalar or array-like int
         If the optional argument `count` is given, only the first
         `count` occurrences are replaced. If negative, replace all.
 
     Returns
     -------
     out : ndarray
-        Output array of str or unicode, depending on input type
+        Output array of str or unicode, depending on input type.
 
     See Also
     --------
@@ -1383,19 +1383,24 @@ def replace(a, old, new, count=None):
     array(['The dwash was fresh', 'Thwas was it'], dtype='<U19')
     
     """
-    a_arr = numpy.asarray(a)
+    a_arr = numpy.asanyarray(a)
+    old = numpy.asanyarray(old)
+    new = numpy.asanyarray(new)
     max_int64 = numpy.iinfo(numpy.int64).max
     counts = numpy._core.umath.count(a_arr, old, 0, max_int64)
     if count is not None:
-        count = numpy.asarray(count)
-        counts = numpy.minimum(counts, count, where=count >= 0, out=counts)
+        count = numpy.asanyarray(count)
+        counts = numpy.where(count < 0, counts,
+                             numpy.minimum(counts, count))
+
     buffersizes = (
         numpy._core.umath.str_len(a_arr)
         + counts * (numpy._core.umath.str_len(new) -
                     numpy._core.umath.str_len(old))
     )
-    max_buffersize = numpy.max(buffersizes)
-    out = numpy.empty(a_arr.shape, dtype=f"{a_arr.dtype.char}{max_buffersize}")
+    # buffersizes is properly broadcast along all inputs.
+    out = numpy.empty_like(a_arr, shape=buffersizes.shape,
+                           dtype=f"{a_arr.dtype.char}{buffersizes.max()}")
     numpy._core.umath._replace(a_arr, old, new, counts, out=out)
     return out