The `.replace()` methods counts the number of matches in order to know the new size. This PR passes on the resulting counts array to the replace ufunc, since if count is 0, the fast path can be taken.  Small example of time gain:
```
In [1]: a = np.array(['ABCDEFGHIJKLMNOPQRSTUVWXYZ' * 100] * 10).view(np.char.chararray)

In [2]: %timeit a.replace('0123456', '')
# This PR
33.2 µs ± 102 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# Current main
45.1 µs ± 16.9 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

p.s. As a side benefit, this also fixes a small bug in the output size calculation, which used the maximum possible number of replacements, not taking into account the requested `count`.

EDIT: the second commit ensures that all inputs properly broadcast against each other for the output, which is easy to do and seems nice to have, since the idea of these functions is to behave like ufuncs. I can separate this one off if controversial.