@mhvk 2024-01-05
The PR ensures that an array of `counts` is passed in, which means that for any element that is already known to have no matches, the fast path will be taken. That way, we can at least somewhat use the knowledge already gained from having done the counting. Furthermore, by adjusting `counts` before doing the length calculating, we assure that the lengths are done right.