@mhvk 2024-01-05
Looking more at broadcasting, currently `count` can broadcast to `a.shape`, but not the reverse, since the result is set to have the same shape as the input. That would be easy to change:
```
count = np.asarray(count)
counts = numpy.where(count >= 0, count, counts)
# And then below
out = numpy.empty(counts.shape, ...)
```
With that, things like this work:
```
np.array('0,0,0,0').view(np.char.chararray).replace('0', '1', count=np.arange(4))
# array(['0,0,0,0', '1,0,0,0', '1,1,0,0', '1,1,1,0'], dtype='<U7')
```
This seems nice enough as an extra addition, especially since presumably for variable-length strings, one can have a ufunc deal with everything to start with. So, I've pushed an update...