@mhvk 2024-01-08
@lysnikolaou - no worries, questions are good (as was clear by you noticing my misunderstanding of negative `count`)

@ngoldbaum - would you be able to do a review of this PR for `.replace()` optimization/broadcasting? (@lysnikolaou already had a pretty good look and approved.)